Tracking in async! (#56)
* Tracking in async!

* Pipelines to include all branches and PRs

* Cleanup and comments

+++ meeshkan/scheduler.py
@@ -5,6 +5,7 @@ from typing import Any, Dict, List, Tuple, Callable, Optional, Union  # For self
 from pathlib import Path
 import uuid
 import os
+import asyncio
 
 import meeshkan.job  # Defines scheduler jobs
 import meeshkan.exceptions
@@ -82,8 +83,9 @@ class Scheduler(object):
         self._running_job = None  # type: Optional[meeshkan.job.Job]
         self._notification_status = dict()  # type: Dict[uuid.UUID, str]
         self._history_by_job = dict()  # type: Dict[uuid.UUID, meeshkan.tracker.TrackerBase]
-        self._job_poller = meeshkan.tracker.TrackingPoller(self.notify_updates)  # type: meeshkan.tracker.TrackingPoller
+        self._job_poller = meeshkan.tracker.TrackingPoller(self.notify_updates)
         self._image_upload = img_upload_func
+        self._event_loop = asyncio.get_event_loop()  # Save the event loop for out-of-thread operations
 
     # Properties and Python magic
 
@@ -125,16 +127,18 @@ class Scheduler(object):
         # Get updates; TODO - vals should be reported once we update schema...
         # pylint: disable=unused-variable
         vals, imgpath = self.query_scalars(job_id, latest_only=True, plot=self._image_upload is not None)
-        download_link = ""
-        if imgpath is not None:
-            try:  # Upload image if we're given an image_upload function...
-                download_link = self._image_upload(imgpath, download_link=True)  # type: ignore
-            except Exception:  # pylint:disable=broad-except
-                LOGGER.error("Could not post image to cloud server!")
-        status = self.notify_listeners(self.submitted_jobs[job_id], download_link, -1, "NA")
-        if imgpath is not None:
-            os.remove(imgpath)
-        return status
+        if vals:  # Only send updates if there exists any updates, doh
+            download_link = ""
+            if imgpath is not None:
+                try:  # Upload image if we're given an image_upload function...
+                    download_link = self._image_upload(imgpath, download_link=True)  # type: ignore
+                except Exception:  # pylint:disable=broad-except
+                    LOGGER.error("Could not post image to cloud server!")
+            status = self.notify_listeners(self.submitted_jobs[job_id], download_link, -1, "NA")
+            if imgpath is not None:
+                os.remove(imgpath)
+            return status
+        return False
 
     def _internal_notifier_loop(self, job: meeshkan.job.Job,
                                 notify_method: Callable[[meeshkan.notifiers.Notifier], None]) -> bool:
@@ -156,16 +160,19 @@ class Scheduler(object):
         if job.stale:
             return
 
+
         self._running_job = job
-        self._job_poller.start(job.id)
+        # Create and schedule a task from the Polling job, so we can cancel it without killing the event loop
+        task = self._event_loop.create_task(self._job_poller.poll(job))  # type: asyncio.Task
         self.notify_listeners_job_start(job)
         try:
             job.launch_and_wait()
         except Exception:  # pylint:disable=broad-except
             LOGGER.exception("Running job failed")
+        finally:
+            task.cancel()
 
         self.notify_listeners_job_end(job)
-        self._job_poller.stop()
         self._running_job = None
         LOGGER.debug("Finished handling job: %s", job)
 
@@ -223,7 +230,7 @@ class Scheduler(object):
             LOGGER.debug("Queue processor started")
 
     def stop(self):
-        self._job_poller.stop()
+        # self._job_poller.stop()
         self._queue_processor.schedule_stop()
         if self._running_job is not None:
             # TODO Add an option to not cancel the currently running job?
